decomp: Compilation fixes
Some files are not actually compiling, see e.g.
https://github.com/Xeeynamo/psyz/actions/runs/33305626358/job/99241552988
which has:

src/libcd/sys.c:148: conflicting types for `CdDataCallback'
src/libsnd/sstickcb.c:3: conflicting types for `SsSetTickCallback'
etc.

For PC build, #ifdef the prototypes that would cause horrible breakage
on PC.

+++ decomp/src/libspu/s_gtm.c
@@ -1,6 +1,6 @@
 #include "libspu_private.h"
 
-int SpuGetTransferMode(void) {
+long SpuGetTransferMode(void) {
     if (_spu_transMode && _spu_transMode == 1) {
         _spu_trans_mode = _spu_transMode;
     } else {
